+++ testWorkspace/tsNode/index.js
@@ -7,5 +7,5 @@ const { double, triple } = require('./double.ts');
 console.log(triple(3));
 console.log(double(21));
 
-require('./matching-line.ts')
+require('./matching-line.ts');
 require('./log.ts');